+++ dart/gui/detail/scenes.hpp
@@ -71,6 +71,7 @@ struct DartScene
   dart::simulation::WorldPtr world;
   std::vector<IkHandle> ikHandles;
   std::vector<dart::gui::Gizmo> gizmos;
+  std::function<std::vector<dart::gui::DebugLabelDescriptor>()> debugLabels;
   std::vector<dart::gui::BodyNodeDragHandle> bodyNodeDragHandles;
   std::vector<dart::gui::KeyboardAction> keyboardActions;
   std::function<void()> preStep;
@@ -129,6 +130,7 @@ struct AppOptions
   bool renderOutputModeExplicit = false;
   std::vector<dart::gui::Panel> panels;
   std::vector<dart::gui::Gizmo> gizmos;
+  std::function<std::vector<dart::gui::DebugLabelDescriptor>()> debugLabels;
   std::vector<dart::gui::InverseKinematicsHandle> ikHandles;
   std::vector<dart::gui::BodyNodeDragHandle> bodyNodeDragHandles;
   std::vector<dart::gui::KeyboardAction> keyboardActions;